Add monopole mode calculation
I add calculations for the monopole mode, which is given as the expected
value of the distance operator with respect to the origin.

The monopole mode is evaluated shortly after the energy and printed in
similarly named files: 'monopolei.dat' for imaginary-time evlution and
'monopole.dat' otherwise.
It is printed together with the energy.
The calculation is performed on the GPU.