Add dipole mode calculation
I add calculations for the dipole mode, which is given as the expected
value of the squared distance operator with respect to the origin.

The dipole mode is evaluated shortly after the monopole and printed in
similarly named files: 'dipolei.dat' for imaginary-time evlution and
'dipole.dat' otherwise. It is printed together with the energy and
monopole. The calculation is performed on the GPU.

This could be combined with the monopole calculations using some kind
of pointer for the kernel function such that only two kernel functions
and print out need to be written. But this is good enough for now at
the very least. Especially since this should not be merged with the
master-branch.